refactor: move helpers.tsx and labels.ts to lib folder
