+++ packages/ai-lab/src/components/index.ts
@@ -2,4 +2,3 @@ export * from './AILabImage';
 export * from './AILabVideo';
 export * from './AILabObjectDetectionUI';
 export * from './SimpleObjectDetectionUI';
-export * from './helpers';
