+++ packages/ai-lab/src/index.ts
@@ -1,2 +1,3 @@
 export * from './components';
 export * from './types';
+export * from './lib';
