+++ packages/ai-lab/src/lib/helpers.tsx
